pyproject.toml
[tool.poetry]
name = "prime-one"
version = "0.1.0"
description = ""
authors = ["hassanp <hassan.nit@gmail.com>"]
readme = "README.md"
packages = [
    { include = "prime_one" }
]

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
langchain-openai = "^0.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
